Updated sprites, maps, music, interface.
Sprites, mainly the coins and the gates, should now be interactive. Upon
collecting a coin, the bonus should increase.
The maps now use the updated version of the brick images - improves
speed.
The music is updated to be more theme friendly. Upon collecting the key,
fast paced music should play.
Added more images to the GUI.